Remove core package
The root package is 'com.bankless'. It should not have a 'core' name because we haven't split the logic into modules.
