+++ src/main/java/com/bankless/app/command/Command.java
@@ -1,0 +1,3 @@
+package com.bankless.app.command;
+
+public class Command{}
