+++ src/main/java/com/bankless/app/controller/QueryController.java
@@ -1,0 +1,20 @@
+package com.bankless.app.controller;
+
+import com.bankless.infrastructure.service.EventService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+public class QueryController {
+
+	@Autowired
+	private EventService eventService;
+
+	@RequestMapping("/status")
+	public String status() {
+		String msg = "status UP!";
+		String result = eventService.publish( msg );
+		return "UP!";
+	}
+
+}
