--- src/main/java/com/bankless/core/app/Application.java
@@ -1,13 +1,0 @@
-package com.bankless.core.app;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication(scanBasePackages={"me.khazaddum.mordor.infrastructure.spring","me.khazaddum.mordor.app"})
-public class Application {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Application.class, args);
-	}
-
-}
