--- src/main/java/com/bankless/core/app/command/Command.java
@@ -1,3 +1,0 @@
-package com.bankless.core.app.command;
-
-public class Command{}
