--- src/main/java/com/bankless/core/domain/model/Greeting.java
@@ -1,21 +1,0 @@
-package com.bankless.core.domain.model;
-
-public class Greeting {
-
-	private final long id;
-	private final String content;
-
-	public Greeting(long id, String content) {
-		this.id = id;
-		this.content = content;
-	}
-
-	public long getId() {
-		return id;
-	}
-
-	public String getContent() {
-		return content;
-	}
-
-}
