--- src/main/java/com/bankless/core/domain/repository/GreetingRepository.java
@@ -1,9 +1,0 @@
-package com.bankless.core.domain.repository;
-
-import com.bankless.core.domain.model.Greeting;
-
-public interface GreetingRepository {
-
-    String upsert(Greeting greeting );
-
-}
