--- src/main/java/com/bankless/core/infrastructure/dto/RequestDto.java
@@ -1,3 +1,0 @@
-package com.bankless.core.infrastructure.dto;
-
-public class RequestDto {}
