--- src/main/java/com/bankless/core/infrastructure/dto/SayHelloRequest.java
@@ -1,12 +1,0 @@
-package com.bankless.core.infrastructure.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class SayHelloRequest extends RequestDto {
-
-    String name;
-
-}
