--- src/main/java/com/bankless/core/infrastructure/dto/SayHelloResponse.java
@@ -1,15 +1,0 @@
-package com.bankless.core.infrastructure.dto;
-
-import com.bankless.core.domain.model.Greeting;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@Setter
-@Getter
-@AllArgsConstructor
-public class SayHelloResponse extends ResponseDto {
-
-    Greeting greeting;
-
-}
