--- src/main/java/com/bankless/core/infrastructure/service/EventService.java
@@ -1,7 +1,0 @@
-package com.bankless.core.infrastructure.service;
-
-public interface EventService {
-
-    public String publish( String msg );
-
-}
