--- src/main/java/com/bankless/core/infrastructure/spring/MyConfiguration.java
@@ -1,16 +1,0 @@
-package com.bankless.core.infrastructure.spring;
-
-import com.bankless.core.infrastructure.service.EventService;
-import com.bankless.core.infrastructure.service.impl.EventServiceConsole;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class MyConfiguration {
-
-    @Bean
-    public EventService eventService() {
-        return new EventServiceConsole();
-    }
-
-}
