+++ src/main/java/com/bankless/domain/model/Greeting.java
@@ -1,0 +1,21 @@
+package com.bankless.domain.model;
+
+public class Greeting {
+
+	private final long id;
+	private final String content;
+
+	public Greeting(long id, String content) {
+		this.id = id;
+		this.content = content;
+	}
+
+	public long getId() {
+		return id;
+	}
+
+	public String getContent() {
+		return content;
+	}
+
+}
