+++ src/main/java/com/bankless/domain/repository/GreetingRepository.java
@@ -1,0 +1,9 @@
+package com.bankless.domain.repository;
+
+import com.bankless.domain.model.Greeting;
+
+public interface GreetingRepository {
+
+    String upsert(Greeting greeting );
+
+}
