+++ src/main/java/com/bankless/infrastructure/dto/ResponseDto.java
@@ -1,0 +1,3 @@
+package com.bankless.infrastructure.dto;
+
+public class ResponseDto {}
