+++ src/main/java/com/bankless/infrastructure/dto/SayHelloRequest.java
@@ -1,0 +1,12 @@
+package com.bankless.infrastructure.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class SayHelloRequest extends RequestDto {
+
+    String name;
+
+}
