+++ src/main/java/com/bankless/infrastructure/dto/SayHelloResponse.java
@@ -1,0 +1,15 @@
+package com.bankless.infrastructure.dto;
+
+import com.bankless.domain.model.Greeting;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Setter
+@Getter
+@AllArgsConstructor
+public class SayHelloResponse extends ResponseDto {
+
+    Greeting greeting;
+
+}
