+++ src/main/java/com/bankless/infrastructure/service/EventService.java
@@ -1,0 +1,7 @@
+package com.bankless.infrastructure.service;
+
+public interface EventService {
+
+    public String publish( String msg );
+
+}
